requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

chromadb==0.5.5

sentence-transformers==4.1.0
transformers==4.52.3

torch==2.6.0
torchvision==0.21.0
accelerate==1.4.0
peft==0.15.2

numpy==1.26.4
pillow==10.4.0
python-docx==1.1.2
PyMuPDF==1.27.2

requests==2.32.3
einops==0.8.0
sentencepiece==0.2.0
tqdm==4.67.1
scikit-learn==1.6.1
huggingface-hub==0.30.2